requirements.txt
scrapfly-sdk[scrapy]
parsel
loguru
gTTS
pandas
openpyxl
python-dotenv
langchain-groq
langchain-core
